\name{Gamlist}
\alias{Gamlist}
\docType{data}
\title{
Gamma list for illustration of the use of the package.
}
\description{
List of numeric vectors of increasing values. List consists of two numeric vectors.
The length of each of these vectors are same as the corresponding vectors in Plist.
For the use of Plist and Gamlist see the documentation of the function retentionBrd.
}
\usage{data("Gamlist")}
\seealso{
\code{\link{Plist}}
}

