\name{Plist}
\alias{Plist}
\docType{data}
\title{
p-value list for illustration of the use of the package.
}
\description{
List of numeric vectors of increasing values each between 0 and 1.
}
\usage{data("Plist")}
\details{
Each numeric vector is the maximum p-values for the associated gamma values. The source data where these p-values are calculated are not reported in this package. For the use of Plist and Gamlist see the documentation of the function retentionBrd.
}
\seealso{
\code{\link{Plist}}
}
